Small edits to SliceStats_M
I removed the units (nm, etc) from the output so that it could be handled as just numbers in excel.  I also changed the delimiter from | to , as that is more standard and therefore makes it few steps to copy and paste into excel. 
I changed the default minimum vacuole slice recognition limit from 600 to 300.  Our math treats it as a radius, but 600 is a typical value for the diameter, 300 for the radius.  I also updated the text of the print statements to make a little clearer what it was asking for.
I updated the input question for the body recognition limit so that it asks the user for the (scaled) radius, instead of the area, to keep it consistent with other user-input measurements.